fix-01: Qi int→long миграция — 18 файлов, 13 задач (QI-C01,CMB-C04/05,PLR-H04/С01,CHR-C01/02,NPC-C01/H04,FRM-C01/M03/M05,DAT-H01,QI-H03,QI-MDL-B)

+++ UnityProject/Assets/Scripts/Combat/DamageCalculator.cs
@@ -68,7 +68,7 @@ namespace CultivationGame.Combat
     public struct DefenderParams
     {
         public int CultivationLevel;
-        public int CurrentQi;
+        public long CurrentQi; // FIX: int→long для Qi > 2.1B на L5+
         public QiDefenseType QiDefense;
         
         public int Agility;
